# Remove memory alloc from encodeDBNode