## Why this should be merged

- [X] Adds compliance tests for `encodeDBNode`
- [X] Adds benchmarks for `encodeDBNode`
- [X] Removed unnecessary memory allocation in `encodeDBNode`


Before:
```
Benchmark_EncodeDBNode/empty_node-12         	27639207	        42.89 ns/op	       2 B/op	       1 allocs/op
Benchmark_EncodeDBNode/has_value-12          	19353940	        64.01 ns/op	       8 B/op	       1 allocs/op
Benchmark_EncodeDBNode/1_child-12            	 5774155	       205.9 ns/op	      49 B/op	       2 allocs/op
Benchmark_EncodeDBNode/2_children-12         	 4090172	       295.1 ns/op	      98 B/op	       2 allocs/op
Benchmark_EncodeDBNode/16_children-12        	  705230	      1706   ns/op	     656 B/op	       2 allocs/op
```

After:
```
Benchmark_EncodeDBNode/empty_node-12         	27004876	        44.51 ns/op	       2 B/op	       1 allocs/op
Benchmark_EncodeDBNode/has_value-12          	20231296	        60.40 ns/op	       8 B/op	       1 allocs/op
Benchmark_EncodeDBNode/1_child-12            	 6382173	       189.9 ns/op	      48 B/op	       1 allocs/op
Benchmark_EncodeDBNode/2_children-12         	 4378076	       280.2 ns/op	      96 B/op	       1 allocs/op
Benchmark_EncodeDBNode/16_children-12        	  716936	      1656   ns/op	     640 B/op	       1 allocs/op
```

## How this works

Uses the same trick from `hashNode` to remove the memory allocation for the keys slice.

## How this was tested

- [X] CI